refactoring: move methods in File helper #599 (#645)
Attempt at resolving issue #599. Currently unsure about testcase, and if I happened to miss some class that used any of the methods found in FileHelper.
