+++ repairnator/repairnator-pipeline/src/test/java/fr/inria/spirals/repairnator/process/utils4tests/ProjectInspectorMocker.java
@@ -1,6 +1,7 @@
 package fr.inria.spirals.repairnator.process.utils4tests;
 
 import fr.inria.spirals.repairnator.BuildToBeInspected;
+import fr.inria.spirals.repairnator.process.files.FileHelper;
 import fr.inria.spirals.repairnator.process.git.GitHelper;
 import fr.inria.spirals.repairnator.process.inspectors.JobStatus;
 import fr.inria.spirals.repairnator.process.inspectors.ProjectInspector;
